requirements.txt
aiogram==2.25.2
pdfplumber
python-docx
pytesseract
doctr
faiss-cpu
peewee
python-dotenv
Pillow
opencv-python
numpy
pandas
openpyxl
psycopg2-binary 